requirements.txt
asgiref==3.8.1
Django==5.2.3
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-taggit==6.1.0
pillow==11.2.1
sqlparse==0.5.3
